Scanner class Prcatice

+++ InputUser/scannerPractice.java
@@ -3,17 +3,17 @@ public class scannerPractice {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
     System.out.println("Enter the first number: ");
-    int Num1 = sc.nextInt();
+    double Num1 = sc.nextDouble();
     System.out.println("Enter the second number: ");
-    int Num2 = sc.nextInt();
+    Double Num2 = sc.nextDouble();
 
     //cler buffer
     sc.nextLine();
 
-    int add = Num1 + Num2;
-    int sub = Num1 - Num2;
-    int multi = Num1 * Num2;
-    int modulus = Num1 / Num2;
+    double add = Num1 + Num2;
+    double sub = Num1 - Num2;
+    double multi = Num1 * Num2;
+    double modulus = Num1 / Num2;
 
     System.out.println("After Addition: " + add);
     System.out.println("After substraction: " + sub);
